+++ airflow/ui/src/components/Clear/ClearAccordion.tsx
@@ -34,7 +34,7 @@ type Props = {
 
 // Table is in memory, pagination and sorting are disabled.
 // TODO: Make a front-end only unconnected table component with client side ordering and pagination
-const ClearRunAccordion = ({ affectedTasks, note, setNote }: Props) => (
+const ClearAccordion = ({ affectedTasks, note, setNote }: Props) => (
   <Accordion.Root collapsible defaultValue={["note"]} multiple={false} variant="enclosed">
     <Accordion.Item key="tasks" value="tasks">
       <Accordion.ItemTrigger>
@@ -98,4 +98,4 @@ const ClearRunAccordion = ({ affectedTasks, note, setNote }: Props) => (
   </Accordion.Root>
 );
 
-export default ClearRunAccordion;
+export default ClearAccordion;
