+++ airflow/ui/src/components/Clear/Run/ClearRunButton.tsx
@@ -20,18 +20,18 @@ import { Box, useDisclosure } from "@chakra-ui/react";
 import { useState } from "react";
 import { FiRefreshCw } from "react-icons/fi";
 
-import type { TaskInstanceCollectionResponse, TaskInstanceResponse } from "openapi/requests/types.gen";
-import { useClearTaskInstances } from "src/queries/useClearTaskInstances";
+import type { DAGRunResponse, TaskInstanceCollectionResponse } from "openapi/requests/types.gen";
+import ActionButton from "src/components/ui/ActionButton";
+import { useClearDagRun } from "src/queries/useClearRun";
 
-import ActionButton from "../ui/ActionButton";
-import ClearTaskInstanceDialog from "./ClearTaskInstanceDialog";
+import ClearRunDialog from "./ClearRunDialog";
 
 type Props = {
-  readonly taskInstance: TaskInstanceResponse;
+  readonly dagRun: DAGRunResponse;
   readonly withText?: boolean;
 };
 
-const ClearTaskInstanceButton = ({ taskInstance, withText = true }: Props) => {
+const ClearRunButton = ({ dagRun, withText = true }: Props) => {
   const { onClose, onOpen, open } = useDisclosure();
 
   const [affectedTasks, setAffectedTasks] = useState<TaskInstanceCollectionResponse>({
@@ -39,10 +39,10 @@ const ClearTaskInstanceButton = ({ taskInstance, withText = true }: Props) => {
     total_entries: 0,
   });
 
-  const dagId = taskInstance.dag_id;
-  const dagRunId = taskInstance.dag_run_id;
+  const dagId = dagRun.dag_id;
+  const dagRunId = dagRun.dag_run_id;
 
-  const { isPending, mutate } = useClearTaskInstances({
+  const { isPending, mutate } = useClearDagRun({
     dagId,
     dagRunId,
     onSuccessConfirm: onClose,
@@ -52,23 +52,23 @@ const ClearTaskInstanceButton = ({ taskInstance, withText = true }: Props) => {
   return (
     <Box>
       <ActionButton
-        actionName="Clear Task Instance"
+        actionName="Clear Dag Run"
         icon={<FiRefreshCw />}
         onClick={onOpen}
-        text="Clear Task Instance"
+        text="Clear Run"
         withText={withText}
       />
 
-      <ClearTaskInstanceDialog
+      <ClearRunDialog
         affectedTasks={affectedTasks}
+        dagRun={dagRun}
         isPending={isPending}
         mutate={mutate}
         onClose={onClose}
         open={open}
-        taskInstance={taskInstance}
       />
     </Box>
   );
 };
 
-export default ClearTaskInstanceButton;
+export default ClearRunButton;
