+++ airflow/ui/src/components/Clear/Run/index.tsx
@@ -17,5 +17,4 @@
  * under the License.
  */
 
-export { default as ClearRunButton } from "./Run";
-export { default as ClearTaskInstanceButton } from "./TaskInstance";
+export { default } from "./ClearRunButton";
